requirements.txt
flask==2.0.1
gunicorn==20.1.0
simplejson==3.17.6
pymongo==4.1.1
dnspython==2.2.1
python-dotenv==0.20.0
jsons==1.6.3
boto3==1.24.70
